# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hope Alive Children Spring Foundation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DonationPaymentMethod =
  (typeof DonationPaymentMethod)[keyof typeof DonationPaymentMethod];

export const DonationPaymentMethod = {
  stripe: "stripe",
  paypal: "paypal",
  korapay: "korapay",
  bank_transfer: "bank_transfer",
} as const;

export type DonationPaymentStatus =
  (typeof DonationPaymentStatus)[keyof typeof DonationPaymentStatus];

export const DonationPaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Donation {
  id: number;
  donorName?: string | null;
  donorEmail?: string | null;
  donorPhone?: string | null;
  amount: number;
  currency: string;
  paymentMethod: DonationPaymentMethod;
  paymentStatus: DonationPaymentStatus;
  transactionId?: string | null;
  purpose?: string | null;
  isAnonymous: boolean;
  createdAt: string;
}

export type CreateDonationBodyPaymentMethod =
  (typeof CreateDonationBodyPaymentMethod)[keyof typeof CreateDonationBodyPaymentMethod];

export const CreateDonationBodyPaymentMethod = {
  stripe: "stripe",
  paypal: "paypal",
  korapay: "korapay",
  bank_transfer: "bank_transfer",
} as const;

export interface CreateDonationBody {
  donorName?: string | null;
  donorEmail?: string | null;
  donorPhone?: string | null;
  amount: number;
  currency: string;
  paymentMethod: CreateDonationBodyPaymentMethod;
  transactionId?: string | null;
  purpose?: string | null;
  isAnonymous?: boolean;
}

export interface VerifyStripeBody {
  sessionId: string;
}

export interface CreateStripeSessionBody {
  amount: number;
  currency: string;
  donorName?: string | null;
  donorEmail?: string | null;
  donorPhone?: string | null;
  purpose?: string | null;
  isAnonymous?: boolean;
}

export interface StripeSessionResponse {
  sessionId: string;
  url: string;
}

export interface DonationListResponse {
  donations: Donation[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  username: string;
  isAdmin: boolean;
}

export type AnalyticsResponseMethodBreakdownItem = {
  method: string;
  count: number;
  total: number;
};

export type AnalyticsResponseMonthlyTotalsItem = {
  month: string;
  total: number;
  count: number;
};

export interface AnalyticsResponse {
  totalDonations: number;
  totalAmount: number;
  totalDonors: number;
  methodBreakdown: AnalyticsResponseMethodBreakdownItem[];
  monthlyTotals: AnalyticsResponseMonthlyTotalsItem[];
  recentDonations: Donation[];
}

export interface Contact {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
  honeypot?: string | null;
}

export interface ContactListResponse {
  contacts: Contact[];
  total: number;
  page: number;
  limit: number;
}

export interface GalleryImage {
  id: number;
  url: string;
  caption?: string | null;
  category?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryImageBody {
  url: string;
  caption?: string | null;
  category?: string | null;
  sortOrder?: number;
}

export interface GalleryListResponse {
  images: GalleryImage[];
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateFaqBody {
  question: string;
  answer: string;
  sortOrder?: number;
}

export interface FaqListResponse {
  faqs: Faq[];
}

export interface ContentItem {
  key: string;
  value: string;
  updatedAt: string;
}

export interface UpdateContentBody {
  value: string;
}

export type AdminListDonationsParams = {
  page?: number;
  limit?: number;
  method?: AdminListDonationsMethod;
};

export type AdminListDonationsMethod =
  (typeof AdminListDonationsMethod)[keyof typeof AdminListDonationsMethod];

export const AdminListDonationsMethod = {
  stripe: "stripe",
  paypal: "paypal",
  korapay: "korapay",
  bank_transfer: "bank_transfer",
} as const;

export type AdminListContactsParams = {
  page?: number;
  limit?: number;
};